переименовал класс, конструктор и создание экземпляра генератора чисел, в класс добавлен метод для агрегации с результатами хранения в Map
